call link_course_to_lead task inside service
